backend/pyproject.toml
[tool.poetry]
name = "streaming-web-backend"
version = "0.1.0"
description = ""
authors = ["NVIDIA"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = ">=3.12, <3.15"
fastapi = "0.115.3"
uvicorn = "^0.30.1"
aerich = "^0.9.1"
tortoise-orm = "^0.25.1"
httpx = "^0.27.0"
pyjwt = "^2.8.0"
pytest-httpx = "^0.30.0"
websockets = "^13.0"
cachetools = "^5.5.0"
asyncache = "^0.3.1"
pytest-mock = "^3.14.0"
cryptography = "^43.0.0"
starlette = "0.40.0"
setuptools = "80.8.0"
h11 = "^0.16.0"
opentelemetry-api = "^1.36.0"
opentelemetry-sdk = "^1.36.0"
opentelemetry-exporter-otlp = "^1.36.0"
opentelemetry-instrumentation-fastapi = "^0.57b0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest-asyncio = "^0.23.7"
pytest = "^8.2.2"
freezegun = "^1.5.1"
asgi-lifespan = "^2.1.0"
respx = "^0.22.0"

[tool.poetry.scripts]
api = "app.main:start"
migrations = "app.main:migrations"
test-metrics = "scripts.metrics:test"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests/**/*.py"]

[tool.black]
line-length = 80

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "app.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
